requirements.txt
pytesseract==0.3.13
Pillow==10.3.0
opencv-python==4.10.0.84
ultralytics==8.3.57
onnxruntime